+++ src/maxLength.js
@@ -1,7 +1,8 @@
 import validator from './validator';
 import { lte } from 'lodash';
 
-export default function maxlengthValidator(max = 0, props = {}) {
+export default function maxLengthValidator(max = 0, props) {
+    props = props || {};
     props.message = props.message || `Length no more than ${max}`;
 
     return validator(