+++ src/minValue.js
@@ -1,7 +1,8 @@
 import validator from './validator';
 import { gte } from 'lodash';
 
-export default function minvalueValidator(min = 0, props = {}) {
+export default function minValueValidator(min = 0, props) {
+    props = props || {};
     props.message = props.message || `At least ${min}`;
 
     return validator(