+++ src/required.js
@@ -1,6 +1,7 @@
 import ValidationResult from './ValidationResult';
 
-export default function requiredValidator(props = {}) {
+export default function requiredValidator(props) {
+    props = props || {};
     props.message = props.message || 'Required';
 
     // Since validator only performs validation for truthy values,