+++ src/validators.js
@@ -1,8 +1,10 @@
 export { default as validator } from './validator';
-export { default as length } from './length';
-export { default as minlength } from './minlength';
-export { default as maxlength } from './maxlength';
-export { default as value } from './value';
-export { default as minvalue } from './minvalue';
-export { default as maxvalue } from './maxvalue';
 export { default as required } from './required';
+
+export { default as value } from './value';
+export { default as minValue } from './minValue';
+export { default as maxValue } from './maxValue';
+
+export { default as length } from './length';
+export { default as minLength } from './minLength';
+export { default as maxLength } from './maxLength';